feat: update

+++ data_loader/retryer_device/normal_device/model.h
@@ -15,8 +15,8 @@ namespace data_loader::retryer_device::normal_device
     {
         std::chrono::nanoseconds base_wait_time;
         std::chrono::nanoseconds max_wait_time;
-        uint32_t exponential_base;
-        uint32_t max_retry_count;
+        double exponential_base;
+        uint64_t max_retry_count;
         std::optional<std::vector<std::string>> retryable_exception_vec;
 
         template <class Reflector>
